requirements.txt
firebase_admin==5.0.1
matplotlib==3.4.2
pyngrok==5.0.5
line-bot-sdk==1.19.0
requests==2.25.1
Flask==2.0.1
bs4==0.0.1